requirements.txt
mysqlclient==2.0.3
dateparser==1.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.6
django-3-jet==1.0.8
django-cors-headers==3.7.0
django-simple-history==2.12.0
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
jet-django==1.1.1
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.7.0
SQLAlchemy==1.4.6
django-import-export==2.5.0
django-environ==0.4.5
